[v0.4.0]Seperate Components
