+++ src/App.tsx
@@ -9,7 +9,6 @@ import Diary from "./routes/Diary.tsx";
 import DiaryDetail from './routes/DiaryDetail.tsx';
 import QuillEditor from "./routes/QuillEditor.tsx";
 import QuillEditorUpdate from "./routes/QuillEditorUpdate.tsx";
-import WebEditor from './routes/WebEditor.js';
 import AuthPage from "./routes/AuthPage.tsx";
 import PrivateRoute from './utils/PrivateRoute.tsx';
 
@@ -32,7 +31,6 @@ const App: React.FC = () => {
   return (
     <BrowserRouter>
         <Routes>
-          <Route path="/webeditor" element={<WebEditor />} />
           <Route element={<Layout/>}>
             <Route path="/" element={<Home />} />
             <Route path="/about" element={<About />} />
