--- src/scss/WebEditor.scss
@@ -1,155 +1,0 @@
-/* 스크롤바 전체 */
-::-webkit-scrollbar {
-    width: 15px;
-}
-/* 스크롤바 트랙 */
-::-webkit-scrollbar-track {
-    background: #f1f1f1;
-    border-radius: 10px;
-}
-/* 스크롤바 핸들 */
-::-webkit-scrollbar-thumb {
-    background-color: #888;
-    border-radius: 10px;
-    border: 3px solid #f1f1f1;
-}
-/* 스크롤바 핸들 호버 */
-::-webkit-scrollbar-thumb:hover {
-    background-color: #555;
-}
-
-.custom-checkbox {
-    width: 15px;
-    height: 15px;
-    transform: scale(1.5); /* 크기를 1.5배로 조정 */
-    margin-left: 10px;
-    top:5px;
-}
-
-h3 {
-    font-size: 15px;
-    font-weight: bold;
-    color: #fff;
-    text-align: center;
-    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.6);
-    letter-spacing: 1px; /* 자간 조정 */
-    border-bottom: 3px solid rgba(255, 255, 255, 0.3); /* 하단 테두리 */
-    background: linear-gradient(135deg, rgba(0, 0, 0, 0.5), rgba(50, 50, 50, 0.5)); /* 배경 그라데이션 */
-    border-radius: 5px;
-}
-
-@mixin padding-margin(
-$padding: 0px,
-$margin: 0px
-){
-    padding: $padding;
-    margin: $margin;
-}
-@mixin outline-setup(
-$box-shadow: 0 4px 10px rgba(0, 0, 0, 0.5), 
-$text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5),
-$border-radius: 5px) {
-    box-shadow: $box-shadow;
-    text-shadow: $text-shadow;
-    border-radius: $border-radius;
-}
-
-.web-editor-inf {
-    @include outline-setup;
-    // @extend %basic;
-    position: absolute;
-    padding: 10px;
-    top: 10px;
-    left: 10px;
-    background-color: rgba(0, 0, 0, 0.7);
-    max-height: 850px;
-    max-width: 500px;
-    overflow-y: auto;
-    overflow-x: hidden;
-    
-    /* 툴팁 설정 */
-    .web-editor-tip {
-        @include outline-setup;
-        @include padding-margin(10px, 0 0 10px 0);
-        font-weight: bold;
-        font-size: 14px;
-        overflow-y: auto;
-        max-height: 300px;
-        background-color: rgba(255, 255, 255, 0.7);
-    }
-
-    /* 조명 설정 */
-    .web-editor-light {
-        @include outline-setup;
-        @include padding-margin(10px, 0 0 10px 0);
-        font-weight: bold;
-        background-color: rgba(255, 255, 255, 0.7);
-    }
-
-    /* 도형 추가 설정 */
-    .web-editor-add-mesh {
-        @include outline-setup;
-        @include padding-margin(10px, 0 0 10px 0);
-        font-weight: bold;
-        background-color: rgba(73, 169, 61, 0.7);
-    }
-
-    /* 도형 수정 설정 */
-    .web-editor-modify-mesh {
-        @include outline-setup;
-        @include padding-margin(10px, 0 0 10px 0);
-        font-weight: bold;
-        background-color: rgba(205, 103, 24, 0.7);
-    }
-
-    /* 생성 도형 목록 설정 */
-    .web-editor-meshes {
-        @include outline-setup;
-        @include padding-margin(10px, 0 0 10px 0);
-        font-weight: bold;
-        background-color: rgba(255, 255, 255, 0.7);
-        max-height: 300px;
-        overflow-y: auto;
-    }
-    
-    /* 업로드 도형 목록 설정 */
-    .web-editor-upload-meshes {
-        @include outline-setup;
-        padding: 10px;
-        font-weight: bold;
-        background-color: rgba(229, 255, 0, 0.7);
-        box-shadow: 0 4px 10px rgba(0, 0, 0, 0.5);
-        max-height: 300px;
-        overflow-y: auto;
-
-        .upload-input {
-            display: none; /* 기본 input 숨기기 */
-        }
-        
-        .upload-label {
-            @include outline-setup;
-            @include padding-margin(10px 20px, 0 5px 5px 0);
-            background: linear-gradient(135deg, #555, #777);
-            border: none;
-            color: white;
-            font-size: 10px;
-            cursor: pointer;
-            transition: transform 0.4s ease, box-shadow 0.4s ease; /* ease 추가 */
-        }
-        .upload-label:hover {
-            transform: translateY(-3px);
-            box-shadow: 0 6px 20px rgba(0, 0, 0, 0.7);
-        }
-        .upload-label:active {
-            transform: translateY(0);
-            box-shadow: 0 2px 5px rgba(0, 0, 0, 0.5);
-        }
-        
-        .web-editor-mini-div {
-            @include outline-setup;
-            font-weight: bold;
-            padding: 10px;
-            background-color: rgba(0, 255, 255, 0.5);
-        }
-    }
-}
